refactor: move keyboard events to directory
